Return user with id on user creation
User create interface (POST) return the user struct with correct
user_id in database.

+++ src/data.go
@@ -1,7 +1,7 @@
 package main
 
 type User struct {
-	UserId   uint32 `json:"user_id" db:"user_id,omitempty" form:"-"`
+	UserId   int64  `json:"user_id" db:"user_id,omitempty" form:"-"`
 	Username string `json:"username" db:"username" form:"username"`
 	Gender   string `json:"gender" db:"gender" form:"gender"`
 }
